update examples, testsuite and bench to account for preferred way to call moved convert tools
