+++ examples/polystyrene_gromacs/polystyrene.py
@@ -13,7 +13,7 @@ import espressopp
 import mpi4py.MPI as MPI
 import logging
 from espressopp import Real3D, Int3D
-from espressopp.tools.convert import gromacs
+from espressopp.tools import gromacs
 from espressopp.tools import decomp
 from espressopp.tools import timers
