+++ testsuite/pi_water/water.py
@@ -15,7 +15,7 @@ import logging
 import copy
 import math
 from espressopp import Real3D, Int3D
-from espressopp.tools.convert import gromacs
+from espressopp.tools import gromacs
 from espressopp.tools import decomp
 from espressopp.tools import timers
 import pathintegral
